// src/app/event/models/registration/RegistrationNotification.ts
export class RegistrationNotification {
    registrantId: string;
    registrantFirstName: string;
    registrantLastName: string;
    registrantImage: string;
    registrantTickets: string;
    registrantTotal: string;
    registrantEmail: string;
    promoterOcode: string;
    eventOrganizerOcode: string;
    eventImageURL: string;
    eventTitle: string;
    eventLocation: string;
    eventId: string;
    eventOwnerOcode: string;
    eventDate: string;

    constructor() {

    }

}